# Optimize RegExpRouter