@yusukebe 2022-03-12
Hi @usualoma 

> I think most middleware is applied to paths like '*' or '/path/to/*' that take no parameters.

I'm thinking about same things. This PR looks good to me!

The score of `RegExpRouter` has been about the same as a score of `TrieRouter` in this case.

![SS](https://user-images.githubusercontent.com/10682/158002630-18c832b5-fe2a-49d4-9883-558fe5d65960.png)

I'll merge it!